fix: enforce no direct Flame imports in app widgets
Move Flame-backed chrome widget implementations into game widget chrome files, keep app widget entrypoints as stable re-exports, and add a repo lint gate to prevent direct package:flame imports under app/lib/widgets.

Made-with: Cursor
