+++ app/lib/features/game/widgets/chrome/flame_region_map_api.dart
@@ -1,0 +1,3 @@
+export 'package:flame/events.dart';
+export 'package:flame/game.dart';
+export 'package:flame/input.dart';
