+++ test/ct_repo_lint_test.dart
@@ -21,6 +21,7 @@ void main() {
       expect(ids, contains('repo.repeated_magic_numbers'));
       expect(ids, contains('repo.function_size'));
       expect(ids, contains('repo.part_unit_size'));
+      expect(ids, contains('repo.no_flame_in_widgets'));
       expect(ids, contains('repo.app_hardcoded_ui_strings'));
       expect(
         rules
