+++ tool/ct_repo_lint_lib.dart
@@ -12,6 +12,7 @@ import 'check_control_flow_nesting_depth.dart';
 import 'check_custom_exceptions.dart';
 import 'check_disallowed_ast_patterns.dart';
 import 'check_function_size.dart';
+import 'check_no_flame_in_widgets.dart';
 import 'check_part_unit_size.dart';
 import 'check_repeated_magic_numbers.dart';
 import 'check_tech_id_constants.dart';
@@ -628,6 +629,8 @@ int? _tryRunDartRuleInProcess({
       return runCheckFunctionSize(repoRoot);
     case 'repo.part_unit_size':
       return runCheckPartUnitSize(repoRoot);
+    case 'repo.no_flame_in_widgets':
+      return runCheckNoFlameInWidgets(repoRoot);
     case 'repo.tech_id_constants':
       return runCheckTechIdConstants(
         repoRoot,
